+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Impl `Clone` for `inspect::*` types;
 - Impl `FromSql`/`ToSql` for `[u8; N]`;
 - Impl `TryFrom<elephantry::Interval>` for `std::time::Duration`;
+- Impl `TryFrom<elephantry::Interval>` for `chrono::Duration`;
 
 ### Fixed
 - `u8` conversion;